Move reducer and change actions
